+++ LakeshoreMarket/src/main/java/service/serviceImpl/CustomerServiceImpl.java
@@ -2,15 +2,6 @@ package main.java.service.serviceImpl;
 
 import java.util.List;
 
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-
 import main.java.DAO.CustomerDAO;
 import main.java.DAO.daoImpl.CustomerDAOImpl;
 import main.java.model.bean.CustomerBean;
@@ -44,7 +35,7 @@ public class CustomerServiceImpl implements CustomerService {
 
 
 	@Override
-	public void delete(long customerId) {
+	public void delete(int customerId) {
 		customerDAO.delete(customerId);
 	}
