+++ LakeshoreMarket/src/main/java/service/serviceImpl/OrderLineServiceImpl.java
@@ -2,15 +2,6 @@ package main.java.service.serviceImpl;
 
 import java.util.List;
 
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-
 import main.java.DAO.OrderLineDAO;
 import main.java.DAO.daoImpl.OrderLineDAOImpl;
 import main.java.model.bean.OrderLineBean;
@@ -32,7 +23,7 @@ public class OrderLineServiceImpl implements OrderLineService{
 	}
 	
 	@Override
-	public OrderLineBean get(long orderLineId) {
+	public OrderLineBean get(int orderLineId) {
 		return ElementUtil.buildOrderLineBean(orderLineDAO.get(orderLineId));
 	}
 	
@@ -43,7 +34,7 @@ public class OrderLineServiceImpl implements OrderLineService{
 	}
 	
 	@Override
-	public void cancel(long orderId) {
+	public void cancel(int orderId) {
 		orderLineDAO.delete(orderId);
 	}
