+++ LakeshoreMarket/src/main/java/service/serviceImpl/PartnerServiceImpl.java
@@ -2,19 +2,9 @@ package main.java.service.serviceImpl;
 
 import java.util.List;
 
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-
 import main.java.DAO.PartnerDAO;
 import main.java.DAO.daoImpl.PartnerDAOImpl;
 import main.java.model.bean.PartnerBean;
-import main.java.model.bean.ProductBean;
 import main.java.service.service.PartnerService;
 import main.java.util.ElementUtil;
 
@@ -49,7 +39,7 @@ public class PartnerServiceImpl implements PartnerService {
 	}
 
 	@Override
-	public void delete(long partnerId) {
+	public void delete(int partnerId) {
 		partnerDAO.delete(partnerId);
 		
 	}
